requirements.txt
streamlit
pandas
matplotlib
pdfplumber
Pillow
google-generativeai
